realtime_stock/requirements.txt
pykis==0.7.0
requests==2.31.0
redis==5.0.1
python-dotenv==1.0.0 